requirements.txt
# alternance-finder — Dépendances Python
pandas
pyyaml
python-dotenv
requests
beautifulsoup4
pytest
pydantic
